added another file for the other class
